New files.
Hi, Alexis... 

I tried to send the new files to you as a zip. 

Please look over these before merging because I don't know if this is correct. 

I believe if you open the Program.cs, that you will find the changes. Basically, what I did was that I changed the Addbooks to a "Book" type instead of a void and then used that as a parameter to feed into the ListBooks to add the book (which is set to default null).